refactor(campaigns): extract bidding-strategy builders to shared module
Pure 1:1 extraction of the bidding-strategy maps, support/required-flag
sets and builder helpers from direct_cli/commands/campaigns.py into a
new direct_cli/_bidding_strategy.py module. Adds a dispatch registry
(CAMPAIGN_TYPE_BUILDERS / register_bidding_strategy_builder /
get_bidding_strategy_builder) keyed on
(campaign_type, operation, branch) so milestone 0.3.13 leaf-PRs
(#361-369, #373) can register a single entry instead of editing the
giant if/elif chain in add()/update().

Behavior is byte-identical; the existing helpers are re-imported in
campaigns.py with their leading-underscore aliases so call sites are
unchanged. Co-located private helpers (_build_package_bidding_strategy,
_build_smart_package_bidding_strategy etc.) remain in campaigns.py
because they are not part of the shared bidding-strategy contract.
